docker config moved to separate plugin
