+++ buildSrc/src/main/groovy/DockerApplication.groovy
@@ -3,7 +3,6 @@ import org.gradle.api.Project
 import org.gradle.api.tasks.Copy
 import org.gradle.api.plugins.ApplicationPlugin
 
-import com.bmuschko.gradle.docker.DockerRemoteApiPlugin
 import com.bmuschko.gradle.docker.tasks.image.Dockerfile
 import com.bmuschko.gradle.docker.tasks.image.DockerBuildImage
 
@@ -11,26 +10,7 @@ class DockerApplication implements Plugin<Project> {
 
 	void apply(Project project) {
 		project.pluginManager.apply(ApplicationPlugin)
-		project.pluginManager.apply(DockerRemoteApiPlugin)
-		
-		def env = System.env
-		project.docker {
-			if(env.containsKey('DOCKER_HOST')) {
-				url = "$env.DOCKER_HOST"
-
-				if(env.containsKey('DOCKER_TLS_VERIFY')) {
-					url = url.replace('tcp', 'https')
-				} else {
-					url = url.replace('tcp', 'http')
-				}
-
-				if(env.containsKey('DOCKER_CERT_PATH')) {
-					certPath = project.file "$env.DOCKER_CERT_PATH"
-				}
-			} else {
-				url = "http://${dockerHost}:2375"
-			}
-		}
+		project.pluginManager.apply(DockerRemoteConfig)
 		
 		def copyTar = project.tasks.create('copyTar', Copy, {
 			def distTar = project.tasks.getByName('distTar')
